implmented user dao -- jdbcTemplate connection guard statements

+++ server/src/main/java/com/server/dao/UserDaoImpl.java
@@ -23,23 +23,25 @@ public class UserDaoImpl extends JdbcDaoSupport implements UserDao {
     }
 
     public void save(User user) {
+        if (getJdbcTemplate() == null) throw new NullPointerException();
         getJdbcTemplate().update(
                 "INSERT INTO public.users (uname, pwd, email) VALUES (?, ?, ?)",
                 user.getName(),
                 user.getPassword(),
                 user.getEmail()
         );
-
     }
 
-    public void delete(String uname) {
+    public void delete(String uname) throws NullPointerException {
+        if (getJdbcTemplate() == null) throw new NullPointerException();
         getJdbcTemplate().update(
                 "DELETE FROM public.users WHERE uname = ?",
                 uname
         );
     }
 
-    public void updateUser(User user) {
+    public void updateUser(User user) throws NullPointerException {
+        if (getJdbcTemplate() == null) throw new NullPointerException();
         getJdbcTemplate().update(
                 "UPDATE public.user " +
                         "SET uname = ?" +
@@ -53,6 +55,7 @@ public class UserDaoImpl extends JdbcDaoSupport implements UserDao {
         User user;
         try {
             String ps = "SELECT * FROM public.users WHERE " + idType + " = ?";
+            if (getJdbcTemplate() == null) throw new NullPointerException();
             user = DataAccessUtils.singleResult( getJdbcTemplate().query(ps,
                     (resultSet, i) -> {
                         User user1 = new User();
